Renaming layer2/decode to layer2/layer2_decode
Seperating GTEST stuff from normal target in makefile
